Map Manifests -> Deployments
